Use genericCompare in RowEquals as well
It was duplicating the logic and genericCompare is more capable.
